Move the sync() function out of gateway to a more generic location Make the update logic actually work by setting the resource version prior to updating the synced object.
